removed unused models, castged unreassignable_seat, clean up in general
