+++ app/Helpers/Helper.php
@@ -2,7 +2,6 @@
 
 namespace App\Helpers;
 use App\Models\Accessory;
-use App\Models\Actionlog;
 use App\Models\Asset;
 use App\Models\AssetModel;
 use App\Models\Component;
@@ -10,7 +9,6 @@ use App\Models\Consumable;
 use App\Models\CustomField;
 use App\Models\CustomFieldset;
 use App\Models\Depreciation;
-use App\Models\LicenseSeat;
 use App\Models\Setting;
 use App\Models\Statuslabel;
 use App\Models\License;
@@ -1531,14 +1529,5 @@ class Helper
         }
         return redirect()->back()->with('error', trans('admin/hardware/message.checkout.error'));
     }
-    public static function unReassignableCount($license)
-    {
 
-        if (!$license->reassignable) {
-            $count = LicenseSeat::where('unreassignable_seat', '=', true)
-                ->where('license_id', '=', $license->id)
-                ->count();
-            return $count;
-        }
-    }
 }
