+++ app/Http/Controllers/Licenses/LicensesController.php
@@ -248,15 +248,8 @@ class LicensesController extends Controller
         }
 
         $users_count = User::where('autoassign_licenses', '1')->count();
-        $total_seats_count = (int) $license->totalSeatsByLicenseID();
-        $available_seats_count = $license->availCount()->count();
-        $unreassignable_seats_count = Helper::unReassignableCount($license);
-        if(!$license->reassignable){
-            $checkedout_seats_count = ($total_seats_count - $available_seats_count - $unreassignable_seats_count );
-        }
-        else {
-            $checkedout_seats_count = ($total_seats_count - $available_seats_count);
-        }
+
+        [$checkedout_seats_count, $total_seats_count, $available_seats_count, $unreassignable_seats_count] = LicenseSeat::usedSeatCount($license);
 
         $this->authorize('view', $license);
         return view('licenses.view', compact('license'))
