+++ app/Http/Transformers/LicensesTransformer.php
@@ -4,6 +4,7 @@ namespace App\Http\Transformers;
 
 use App\Helpers\Helper;
 use App\Models\License;
+use App\Models\LicenseSeat;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Database\Eloquent\Collection;
 
@@ -37,7 +38,7 @@ class  LicensesTransformer
             'notes' => Helper::parseEscapedMarkedownInline($license->notes),
             'expiration_date' => Helper::getFormattedDateObject($license->expiration_date, 'date'),
             'seats' => (int) $license->seats,
-            'free_seats_count' => (int) $license->free_seats_count - Helper::unReassignableCount($license),
+            'free_seats_count' => (int) $license->free_seats_count - LicenseSeat::unReassignableCount($license),
             'min_amt' => ($license->min_amt) ? (int) ($license->min_amt) : null,
             'license_name' =>  ($license->license_name) ? e($license->license_name) : null,
             'license_email' => ($license->license_email) ? e($license->license_email) : null,
